refactor: rename component files with their real name
