+++ src/renderer/modules/components/Spinner.tsx
